Created BaseRabbit controller
